+++ src/index.js
@@ -81,3 +81,20 @@ export { Labels } from "./components/Labels/Labels";
  */
 
 export { Breadcrumb } from "./components/Breadcrumb/Breadcrumb";
+
+/**
+ *Text Input Field
+ */
+
+export { TextField } from "./components/TextField/TextField";
+
+/*
+ * FormError
+ */
+
+export { FormError } from "./components/FormError/FormError";
+
+/**
+ * Form Label
+ */
+export { FormLabel } from "./components/FormLabel/FormLabel";